refactor input group
