+++ resources/js/admin/Components/Form/Label.vue
@@ -3,12 +3,14 @@ defineProps({
     value: {
         type: String,
     },
+    required: Boolean
 });
 </script>
 
 <template>
-    <label class="block font-medium text-sm text-gray-700">
+    <label class="block text-sm font-medium text-gray-700">
         <span v-if="value">{{ value }}</span>
+        <span v-if="required" class="text-red-600">&#160;*</span>
         <span v-else><slot /></span>
     </label>
 </template>
